Add new tile properties; add new game day constants; fix click through on selection scroll; add new gamestate data points;

+++ core/src/com/twojeremys/awesometower/screen/menu/StatusMenu.java
@@ -89,7 +89,7 @@ public class StatusMenu {
 		this.clockLabel.setText(clockValue);
 	}
 	
-	public void setCoins(int coinValue){
+	public void setCoins(float coinValue){
 		this.coinLabel.setText(String.valueOf(coinValue));
 	}
